Added coverage for Recieving System and Measure Evaluation. App.tsx coverage above 80%
