+++ src/constants/Constants.ts
@@ -1,46 +1,46 @@
 import { Server } from '../models/Server';
 
 export class Constants {
-    public static missingProperty = 'Missing required property: {0}';
-    public static unreachableURL = 'The target URL is unreachable: '
-    public static fetchError = 'Using {0} to retrieve {1} caused: {2}';
-    public static error_receivingSystemServer = 'Please select a Receiving System server to use';
-    public static error_measureEvaluationServer = 'Please select a Measure Evaluation server to use';
-    public static error_selectTestServer = 'Please select a Test Server to use';
-    public static error_selectMeasure = 'Please select a Measure to evaluate';
-    public static error_selectMeasureDR = 'Please select a Measure to get the data requirements for';
-    public static error_selectDataRepository = 'Please select a Data Repository server to use';
-    public static error_selectMeasureDataCollection = 'Please select a Measure to collect the data for';
-    public static error_generateMeasureReport = 'Please evaluate a Measure to generate a MeasureReport to post';
+    public static readonly missingProperty = 'Missing required property: {0}';
+    public static readonly unreachableURL = 'The target URL is unreachable: '
+    public static readonly fetchError = 'Using {0} to retrieve {1} caused: {2}';
+    public static readonly error_receivingSystemServer = 'Please select a Receiving System server to use';
+    public static readonly error_measureEvaluationServer = 'Please select a Measure Evaluation server to use';
+    public static readonly error_selectTestServer = 'Please select a Test Server to use';
+    public static readonly error_selectMeasure = 'Please select a Measure to evaluate';
+    public static readonly error_selectMeasureDR = 'Please select a Measure to get the data requirements for';
+    public static readonly error_selectDataRepository = 'Please select a Data Repository server to use';
+    public static readonly error_selectMeasureDataCollection = 'Please select a Measure to collect the data for';
+    public static readonly error_generateMeasureReport = 'Please evaluate a Measure to generate a MeasureReport to post';
     
-    public static error_url = 'Please provide a valid URL: ';
-    public static error_urlStartsWith = 'Must start with http:// or https://';
-    public static error_urlEndsWith = 'Must end with /';
+    public static readonly error_url = 'Please provide a valid URL: ';
+    public static readonly error_urlStartsWith = 'Must start with http:// or https://';
+    public static readonly error_urlEndsWith = 'Must end with /';
     
-    public static error_selectReceivingSystemServer = 'Please select a Data Repository server to use';
-    public static error_selectMeasureToSubmit = 'Please select a Measure to collect the data for';
+    public static readonly error_selectReceivingSystemServer = 'Please select a Data Repository server to use';
+    public static readonly error_selectMeasureToSubmit = 'Please select a Measure to collect the data for';
 
-    public static evaluateMeasureFetchURL = '{0}Measure/{1}/$evaluate-measure?periodStart={2}&periodEnd={3}&reportType=subject-list';
-    public static evaluateMeasureWithPatientFetchURL = '{0}Measure/{1}/$evaluate-measure?subject={2}&periodStart={3}&periodEnd={4}';
+    public static readonly evaluateMeasureFetchURL = '{0}Measure/{1}/$evaluate-measure?periodStart={2}&periodEnd={3}&reportType=subject-list';
+    public static readonly evaluateMeasureWithPatientFetchURL = '{0}Measure/{1}/$evaluate-measure?subject={2}&periodStart={3}&periodEnd={4}';
 
-    public static defaultStartDate = '2025-01-01';
-    public static defaultEndDate = '2025-12-31';
+    public static readonly defaultStartDate = '2025-01-01';
+    public static readonly defaultEndDate = '2025-12-31';
 
-    public static patientUrlEnding = 'Patient?_count=';
-    public static patientTotalCountUrlEnding = 'Patient?_summary=count';
+    public static readonly patientUrlEnding = 'Patient?_count=';
+    public static readonly patientTotalCountUrlEnding = 'Patient?_summary=count';
     
-    public static measureUrlEnding = 'Measure?_count=200';
+    public static readonly measureUrlEnding = 'Measure?_count=200';
 
-    public static error_selectKnowledgeRepository = 'Please select a Knowledge Repository server to use';
+    public static readonly error_selectKnowledgeRepository = 'Please select a Knowledge Repository server to use';
 
-    public static dataSubmitted = 'Data Submitted';
-    public static submitDataFetchDataError = 'This function has not been implemented into SubmitDataFetch.  Use submitData instead.';
+    public static readonly dataSubmitted = 'Data Submitted';
+    public static readonly submitDataFetchDataError = 'This function has not been implemented into SubmitDataFetch.  Use submitData instead.';
 
-    public static measurePosted = 'Measure Posted';
-    public static measurePostedFetchDataError = 'There was an error posting the measure';
+    public static readonly measurePosted = 'Measure Posted';
+    public static readonly measurePostedFetchDataError = 'There was an error posting the measure';
 
     //testing purposes:
-    public static serverTestData: Server[] = [
+    public static readonly serverTestData: Server[] = [
         {
             id: 'ec2345-1',
             baseUrl: 'http://localhost:8080/1/',
@@ -83,6 +83,6 @@ export class Constants {
         }
     ];
 
-    public static testOauthServer = Constants.serverTestData[3];
-    static submitPostTestBody = `{'prop1': 'val1', 'prop2': 'val2'}`
+    public static readonly testOauthServer = Constants.serverTestData[3];
+    static readonly submitPostTestBody = `{'prop1': 'val1', 'prop2': 'val2'}`
 }
